+++ src/dznpy/adv_shell/port_selection.py
@@ -1,5 +1,6 @@
 """
-Module containing classes for selecting ports and their runtime semantics of Dezyne systems and components.
+Module containing classes for selecting ports and their runtime semantics of Dezyne systems
+and components.
 
 Copyright (c) 2023-2024 Michael van de Ven <michael@ftr-ict.com>
 This is free software, released under the MIT License. Refer to dznpy/LICENSE.